requirements.txt
crewai
openai
python-dotenv
duckduckgo-search